downgrade mdx loader
