+++ website/.alexrc.js
@@ -1,0 +1,15 @@
+// Use a "maybe" level of profanity instead of the default "unlikely".
+exports.profanitySureness = 1
+
+exports.allow = [
+  "simple",
+  "special",
+  "invalid",
+  "he-she",
+  "her-him",
+  "herself-himself",
+  "obvious",
+  "host",
+  "dad-mom",
+  "host-hostess",
+]
