+++ website/app/core/hooks/useIsDocsIndex.js
@@ -1,4 +1,4 @@
-import {useRouter} from "blitz"
+import {useRouter} from "next/router"
 
 export function useIsDocsIndex() {
   return useRouter().pathname === "/docs"
